# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trinity AI Engineering OS API
servers:
  - url: /api
    description: Base API path
security:
  - TrinityApiKey: []
  - BearerAuth: []
...
paths:
...
  /designs/cad:
    post:
      operationId: createCadDesign
      tags: [designs]
      summary: Create and validate a CAD design job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadDesignRequest"
      responses:
        "201":
          description: CAD design job with persisted artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignJobResponse"
  /designs/pcb:
    post:
      operationId: createPcbDesign
      tags: [designs]
      summary: Create and validate a PCB design job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PcbDesignRequest"
      responses:
        "201":
          description: PCB design job with persisted artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignJobResponse"
  /design-jobs/{job_id}:
    get:
      operationId: getDesignJob
      tags: [designs]
      summary: Get a persisted design job
      parameters:
        - name: job_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Design job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignJob"
  /design-jobs/{job_id}/validation:
    get:
      operationId: getDesignJobValidation
      tags: [designs]
      summary: Get design validation details
      parameters:
        - name: job_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Validation report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationReport"
  /artifacts/{artifact_id}:
    get:
      operationId: downloadArtifact
      tags: [designs]
      summary: Download a generated artifact
      parameters:
        - name: artifact_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Binary artifact content
          content:
            application/octet-stream: {}
...
  /firmware/jobs:
    post:
      operationId: createFirmwareJob
      tags: [firmware]
      summary: Generate a target-specific firmware project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirmwareRequest"
      responses:
        "201":
          description: Firmware project job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirmwareJobResponse"
  /firmware/jobs/{job_id}:
    get:
      operationId: getFirmwareJob
      tags: [firmware]
      summary: Get a firmware job
      parameters:
        - name: job_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Firmware job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirmwareJobResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Trinity-HMAC
      description: Short-lived HMAC bearer token issued by the account login endpoint.
    TrinityApiKey:
      type: apiKey
      in: header
      name: X-Trinity-Api-Key
      description: Optional in development; required for protected production API endpoints when TRINITY_API_KEY is configured.
  schemas:
...
    ValidationCheck:
      type: object
      required: [name, status, message]
      properties:
        name: {type: string}
        status: {type: string}
        message: {type: string}
        details: {type: object}
    ValidationReport:
      type: object
      required: [status, checks]
      properties:
        status: {type: string}
        checks:
          type: array
          items: {$ref: "#/components/schemas/ValidationCheck"}
        tool: {type: [string, "null"]}
        tool_version: {type: [string, "null"]}
    Artifact:
      type: object
      required: [id, kind, filename, mime_type, size_bytes, sha256, download_url]
      properties:
        id: {type: string}
        kind: {type: string}
        filename: {type: string}
        mime_type: {type: string}
        size_bytes: {type: integer}
        sha256: {type: string}
        download_url: {type: string}
        preview_url: {type: [string, "null"]}
    CadDesignRequest:
      type: object
      required: [description]
      properties:
        description: {type: string}
        parameters: {type: object}
        output_formats:
          type: array
          items: {type: string}
    PcbDesignRequest:
      type: object
      required: [description]
      properties:
        description: {type: string}
        components:
          type: array
          items: {type: string}
        spec: {type: [object, "null"]}
        outputs:
          type: array
          items: {type: string}
    DesignJob:
      type: object
      required: [job_id, engine, status, request_hash, artifacts, validation, created_at, updated_at]
      properties:
        job_id: {type: string}
        engine: {type: string}
        status: {type: string}
        request_hash: {type: string}
        spec: {type: [object, "null"]}
        artifacts:
          type: array
          items: {$ref: "#/components/schemas/Artifact"}
        validation: {$ref: "#/components/schemas/ValidationReport"}
        questions:
          type: array
          items: {type: string}
        assumptions:
          type: array
          items: {type: string}
        error: {type: [string, "null"]}
        created_at: {type: string}
        updated_at: {type: string}
    DesignJobResponse:
      type: object
      required: [job, legacy]
      properties:
        job: {$ref: "#/components/schemas/DesignJob"}
        legacy: {type: object}
...
    PinAssignment:
      type: object
      required: [name, pin, function]
      properties:
        name: {type: string}
        pin: {type: string}
        function: {type: string}
        active_high: {type: boolean}
    PeripheralSpec:
      type: object
      required: [name, kind]
      properties:
        name: {type: string}
        kind: {type: string}
        bus: {type: [string, "null"]}
        pins:
          type: array
          items: {type: string}
        options: {type: object}
    FirmwareRequest:
      type: object
      required: [description]
      properties:
        description: {type: string}
        target_id: {type: [string, "null"]}
        project_name: {type: string}
        features:
          type: array
          items: {type: string}
        pins:
          type: array
          items: {$ref: "#/components/schemas/PinAssignment"}
        peripherals:
          type: array
          items: {$ref: "#/components/schemas/PeripheralSpec"}
        include_tests: {type: boolean}
        safety_mode: {type: string}
    FirmwareArtifact:
      type: object
      required: [id, filename, kind, mime_type, size_bytes, sha256, download_url]
      properties:
        id: {type: string}
        filename: {type: string}
        kind: {type: string}
        mime_type: {type: string}
        size_bytes: {type: integer}
        sha256: {type: string}
        download_url: {type: string}
    FirmwareCheck:
      type: object
      required: [name, status, message]
      properties:
        name: {type: string}
        status: {type: string}
        message: {type: string}
        details: {type: object}
    FirmwareValidation:
      type: object
      required: [status, checks]
      properties:
        status: {type: string}
        checks:
          type: array
          items: {$ref: "#/components/schemas/FirmwareCheck"}
        toolchain: {type: [string, "null"]}
    FirmwareSpec:
      type: object
      required: [target_id, project_name, description, features, pins, peripherals, include_tests, safety_mode]
      properties:
        target_id: {type: string}
        project_name: {type: string}
        description: {type: string}
        language: {type: [string, "null"]}
        framework: {type: [string, "null"]}
        features:
          type: array
          items: {type: string}
        pins:
          type: array
          items: {$ref: "#/components/schemas/PinAssignment"}
        peripherals:
          type: array
          items: {$ref: "#/components/schemas/PeripheralSpec"}
        include_tests: {type: boolean}
        safety_mode: {type: string}
    FirmwareJob:
      type: object
      required: [job_id, engine, spec, status, validation, artifacts, created_at, updated_at]
      properties:
        job_id: {type: string}
        engine: {type: string}
        target: {type: [object, "null"]}
        spec: {$ref: "#/components/schemas/FirmwareSpec"}
        status: {type: string}
        artifacts:
          type: array
          items: {$ref: "#/components/schemas/FirmwareArtifact"}
        validation: {$ref: "#/components/schemas/FirmwareValidation"}
        questions:
          type: array
          items: {type: string}
        assumptions:
          type: array
          items: {type: string}
        error: {type: [string, "null"]}
        created_at: {type: string}
        updated_at: {type: string}
    FirmwareJobResponse:
      type: object
      required: [job]
      properties:
        job: {$ref: "#/components/schemas/FirmwareJob"}
...