# lib/api-zod/src/generated/types/firmwareCheck.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trinity AI Engineering OS API
 * OpenAPI spec version: 0.1.0
 */
import type { FirmwareCheckDetails } from './firmwareCheckDetails';

export interface FirmwareCheck {
  name: string;
  status: string;
  message: string;
  details?: FirmwareCheckDetails;
}