# lib/api-zod/src/generated/types/firmwareJob.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trinity AI Engineering OS API
 * OpenAPI spec version: 0.1.0
 */
import type { FirmwareArtifact } from './firmwareArtifact';
import type { FirmwareJobTarget } from './firmwareJobTarget';
import type { FirmwareSpec } from './firmwareSpec';
import type { FirmwareValidation } from './firmwareValidation';

export interface FirmwareJob {
  job_id: string;
  engine: string;
  /** @nullable */
  target?: FirmwareJobTarget;
  spec: FirmwareSpec;
  status: string;
  artifacts: FirmwareArtifact[];
  validation: FirmwareValidation;
  questions?: string[];
  assumptions?: string[];
  /** @nullable */
  error?: string | null;
  created_at: string;
  updated_at: string;
}