# lib/api-zod/src/generated/types/pcbDesignRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trinity AI Engineering OS API
 * OpenAPI spec version: 0.1.0
 */
import type { PcbDesignRequestSpec } from './pcbDesignRequestSpec';

export interface PcbDesignRequest {
  description: string;
  components?: string[];
  /** @nullable */
  spec?: PcbDesignRequestSpec;
  outputs?: string[];
}