# lib/api-zod/src/generated/types/validationReport.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Trinity AI Engineering OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationCheck } from './validationCheck';

export interface ValidationReport {
  status: string;
  checks: ValidationCheck[];
  /** @nullable */
  tool?: string | null;
  /** @nullable */
  tool_version?: string | null;
}